Change .gitmodules to improve how forking behaves
Before this commit, recursively cloning a fork of this repository will fail unless one created a fork of each submodule as well.
This commit changes this so that recursively cloning the fork will clone the official nerficg-project versions of each submodule.
If you want to work with your own fork of a submodule you can clone non-recursively, edit the `.gitmodules` file, and then use `git submodule update --init --recursive`.